requirements.txt
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
gitpython>=3.1.0
pyyaml>=6.0
rich>=13.0.0

